+++ crates/client/tests/test_program_client.rs
@@ -1,4 +1,5 @@
 use anyhow::Error;
+use cargo_metadata::camino;
 use fehler::throws;
 use pretty_assertions::assert_str_eq;
 
@@ -23,13 +24,17 @@ pub async fn generate_program_client() {
     let program_idl =
         trdelnik_client::idl::parse_to_idl_program("escrow".to_owned(), expanded_escrow)?;
 
-    let idl = trdelnik_client::idl::Idl {
-        programs: vec![program_idl],
-    };
+    let program_data = vec![(
+        String::default(),
+        camino::Utf8PathBuf::default(),
+        program_idl,
+    )];
 
     let use_modules: Vec<syn::ItemUse> = vec![syn::parse_quote! { use trdelnik_client::*; }];
-    let client_code =
-        trdelnik_client::program_client_generator::generate_source_code(&idl, &use_modules);
+    let client_code = trdelnik_client::program_client_generator::generate_source_code(
+        &program_data,
+        &use_modules,
+    );
     let client_code = trdelnik_client::Commander::format_program_code(&client_code).await?;
 
     assert_str_eq!(client_code, expected_client_code);